Separate isolate logic into new file
