+++ pkg/clusteragent/clusterchecks/dispatcher_rebalance.go
@@ -448,42 +448,3 @@ func rebalanceIsWorthIt(currentDistribution checksDistribution, proposedDistribu
 	maxStdDevAccepted := currentDistribution.utilizationStdDev() * ((100 - float64(minPercImprovement)) / 100)
 	return proposedDistribution.utilizationStdDev() < maxStdDevAccepted
 }
-
-func (d *dispatcher) isolateCheck(isolateCheckID string) types.IsolateResponse {
-	currentDistribution := d.currentDistribution()
-	isolateNode := currentDistribution.runnerForCheck(isolateCheckID)
-	if isolateNode == "" {
-		return types.IsolateResponse{
-			CheckID:   isolateCheckID,
-			CheckNode: "",
-			Result:    false,
-			Reason:    "Unable to find check",
-		}
-	}
-
-	proposedDistribution := newChecksDistribution(currentDistribution.runnerWorkers())
-
-	for _, checkID := range currentDistribution.checksSortedByWorkersNeeded() {
-		if checkID == isolateCheckID {
-			// Keep the check to be isolated on its current runner
-			continue
-		}
-
-		workersNeededForCheck := currentDistribution.workersNeededForCheck(checkID)
-		runnerForCheck := currentDistribution.runnerForCheck(checkID)
-
-		proposedDistribution.addToLeastBusy(
-			checkID,
-			workersNeededForCheck,
-			runnerForCheck,
-			isolateNode,
-		)
-	}
-
-	d.applyDistribution(proposedDistribution, currentDistribution)
-	return types.IsolateResponse{
-		CheckID:   isolateCheckID,
-		CheckNode: isolateNode,
-		Result:    true,
-	}
-}
