+++ pkg/clusteragent/clusterchecks/dispatcher_rebalance_test.go
@@ -1628,81 +1628,3 @@ func TestRebalanceIsWorthIt(t *testing.T) {
 
 	assert.True(t, rebalanceIsWorthIt(currentDistribution, proposedDistribution, 10))
 }
-
-func TestIsolateCheck(t *testing.T) {
-	testDispatcher := newDispatcher()
-	testDispatcher.store.nodes["A"] = newNodeStore("A", "")
-	testDispatcher.store.nodes["A"].workers = config.DefaultNumWorkers
-	testDispatcher.store.nodes["B"] = newNodeStore("B", "")
-	testDispatcher.store.nodes["B"].workers = config.DefaultNumWorkers
-
-	testDispatcher.store.nodes["A"].clcRunnerStats = map[string]types.CLCRunnerStats{
-		"checkA0": {
-			AverageExecutionTime: 50,
-			MetricSamples:        10,
-			IsClusterCheck:       true,
-		},
-		"checkA1": {
-			AverageExecutionTime: 20,
-			MetricSamples:        10,
-			IsClusterCheck:       true,
-		},
-		"checkA2": {
-			AverageExecutionTime: 100,
-			MetricSamples:        10,
-			IsClusterCheck:       true,
-		},
-	}
-
-	testDispatcher.store.nodes["B"].clcRunnerStats = map[string]types.CLCRunnerStats{
-		"checkB0": {
-			AverageExecutionTime: 50,
-			MetricSamples:        10,
-			IsClusterCheck:       true,
-		},
-		"checkB1": {
-			AverageExecutionTime: 20,
-			MetricSamples:        10,
-			IsClusterCheck:       true,
-		},
-		"checkB2": {
-			AverageExecutionTime: 100,
-			MetricSamples:        10,
-			IsClusterCheck:       true,
-		},
-	}
-	testDispatcher.store.idToDigest = map[checkid.ID]string{
-		"checkA0": "digestA0",
-		"checkA1": "digestA1",
-		"checkA2": "digestA2",
-		"checkB0": "digestB0",
-		"checkB1": "digestB1",
-		"checkB2": "digestB2",
-	}
-	testDispatcher.store.digestToConfig = map[string]integration.Config{
-		"digestA0": {},
-		"digestA1": {},
-		"digestA2": {},
-		"digestB0": {},
-		"digestB1": {},
-		"digestB2": {},
-	}
-	testDispatcher.store.digestToNode = map[string]string{
-		"digestA0": "A",
-		"digestA1": "A",
-		"digestA2": "A",
-		"digestB0": "B",
-		"digestB1": "B",
-		"digestB2": "B",
-	}
-
-	// rebalance checks
-	testDispatcher.isolateCheck("checkA2")
-
-	assert.EqualValues(t, len(testDispatcher.store.nodes["A"].clcRunnerStats), 1)
-	assert.EqualValues(t, len(testDispatcher.store.nodes["B"].clcRunnerStats), 5)
-	_, containsIsolatedCheck := testDispatcher.store.nodes["A"].clcRunnerStats["checkA2"]
-	assert.EqualValues(t, containsIsolatedCheck, true)
-
-	requireNotLocked(t, testDispatcher.store)
-}
